backend/core/requirements.txt
fastapi
uvicorn
pydantic
email-validator>=2.0.0
pandas
scikit-learn
numpy
google-api-python-client
pytrends
requests
python-dotenv
yt-dlp
openai
google-generativeai
prophet
statsmodels
tensorflow
scipy
yake
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt>=4.0.0,<5
PyJWT>=2.8.0
langgraph
langchain
langchain-core
langchain-openai
